DocumentController -> RestDocumentListController
- 요약 탭과 전체보기 탭의 목록 불러오기 메서드 일원화
