+++ CMSApp-main/src/main/resources/mybatis/DocumentMapper.xml
@@ -114,9 +114,13 @@
 		left join mastercode m
 		on s.status_code = m.status_code
 		
-		<where>
+		left join trash t
+        on d.document_idx = t.document_idx
+        
+        where trash_idx is null
+		
 		<if test="status_code != 900">
-			s.status_code=#{status_code}
+			and s.status_code=#{status_code}
 		</if>
 		
 		<if test="end_date != null">
@@ -126,7 +130,6 @@
 		<if test="start_date != null">
 			and s.regdate <![CDATA[>=]]> #{start_date}
 		</if>
-		</where>
 	</select>
 	
 	
@@ -147,6 +150,11 @@
         ) s
         
 		on d.document_idx = s.document_idx
+		
+		left join trash t
+        on d.document_idx = t.document_idx
+        
+        where trash_idx is null
         
         group by status_code
         having status_code=#{status_code};
